backend: cleaned asserts
